feat: 비즈니스 로직 예외사항 처리

+++ src/main/java/kit/prolog/service/CategoryService.java
@@ -2,9 +2,9 @@ package kit.prolog.service;
 
 import kit.prolog.controller.CategoryController;
 import kit.prolog.domain.Category;
-import kit.prolog.dto.CategoryDto;
 import kit.prolog.dto.CategoryInfoDto;
 import kit.prolog.repository.jpa.CategoryRepository;
+import kit.prolog.repository.jpa.PostRepository;
 import kit.prolog.repository.jpa.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
@@ -20,12 +20,21 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final UserRepository userRepository;
+    private final PostRepository postRepository;
 
     public void insertCategory(CategoryController.CategoryFormDto categoryFormDto, Long userId) {
+        Category upperCategory = categoryRepository.findById(categoryFormDto.getUpperId()).get();
+        if (upperCategory.getUser().getId() != userId)
+            throw new AccessDeniedException("");
+
+        if (upperCategory.getUpperCategory() != null
+                && upperCategory.getUpperCategory().getUpperCategory() != null)
+            throw new IllegalArgumentException();
+
         Category category = Category.builder()
                 .name(categoryFormDto.getName())
                 .user(userRepository.findById(userId).get())
-                .upperCategory(categoryRepository.findById(categoryFormDto.getUpperId()).get())
+                .upperCategory(upperCategory)
                 .build();
         categoryRepository.save(category);
     }
@@ -34,9 +43,15 @@ public class CategoryService {
         Category category = categoryRepository.findById(categoryId).get();
         if (category.getUser().getId() != userId)
             throw new AccessDeniedException("");
+        if (category.getUpperCategory() == null)
+            throw new AccessDeniedException("");
+
+        Category upperCategory = categoryRepository.findById(categoryFormDto.getUpperId()).get();
+        if (upperCategory.getUser().getId() != userId)
+            throw new AccessDeniedException("");
 
         category.setName(categoryFormDto.getName());
-        category.setUpperCategory(categoryRepository.findById(categoryFormDto.getUpperId()).get());
+        category.setUpperCategory(upperCategory);
         categoryRepository.save(category);
     }
 
@@ -44,7 +59,16 @@ public class CategoryService {
         Category category = categoryRepository.findById(categoryId).get();
         if (category.getUser().getId() != userId)
             throw new AccessDeniedException("");
+        if (category.getUpperCategory() == null)
+            throw new AccessDeniedException("");
 
+        Category newCategory = category.getUpperCategory();
+        postRepository.updatePostCategory(category, newCategory);
+        categoryRepository.findByUpperCategory(category)
+                .stream()
+                .forEach(lowerCategory -> postRepository.updatePostCategory(lowerCategory, newCategory));
+
+        categoryRepository.deleteByUpperCategory(category);
         categoryRepository.delete(category);
     }
